vildrose_core::word

Struct Tryte

pub struct Tryte(/* private fields */);

A fixed-width balanced ternary word type with a width of $width trits.

Implementations

impl Tryte

Construction

pub const fn new(t: [Trit; 9]) -> Self

Creates a new word filled with zeros

pub const fn from_trits(trits: [Trit; 9]) -> Self

Creates a new word from an array of trits

pub const fn zero() -> Self

Creates a new word filled with zeros

pub fn trit(&self, i: usize) -> Trit

Returns the trit at the given index, panicking if the index is out of bounds.

impl Tryte

Sign and magnitude

pub fn negate(&self) -> Self

Returns the inverted (negated) form of the word

pub fn abs(&self) -> Self

Returns the absolute (no negatives) form of the word

pub fn sign(&self) -> Trit

Returns the sign (whether it’s negative, positive or zero) of the word

impl Tryte

Tritwise logic

pub fn tmin(&self, rhs: Self) -> Self

Returns the minimum of two words, element-wise

pub fn tmax(&self, rhs: Self) -> Self

Returns the maximum of two words, element-wise

pub fn tnot(&self) -> Self

Returns the tritwise negation of the word

pub fn tclip(&self) -> Self

Returns the tritwise clipping of the word

pub fn tsign(&self) -> Self

Returns the tritwise sign of the word

pub fn tconsensus(&self, rhs: Self) -> Self

Returns the tritwise consensus of two words

impl Tryte

Shifts

pub fn tshl(&self, n: usize) -> Self

Trit shift left by n positions, filling with zeros

pub fn tshr(&self, n: usize) -> Self

Trit shift right by n positions, filling with the sign of the word

pub fn tlshr(&self, n: usize) -> Self

Trit logical shift right by n positions, filling with zeros